+++ app/src/main/java/com/example/luckyspinner/fragments/SpinnerListFragment.kt
@@ -27,9 +27,11 @@ import com.example.luckyspinner.databinding.EditDialogBinding
 import com.example.luckyspinner.databinding.FragmentSpinnerListBinding
 import com.example.luckyspinner.interfaces.OnEditClickListener
 import com.example.luckyspinner.models.Spinner
+import com.example.luckyspinner.util.Constants.EMPTY_STRING
 import com.example.luckyspinner.util.Constants.ID_CHANNEL_KEY
 import com.example.luckyspinner.util.Constants.ID_SPINNER_KEY
 import com.example.luckyspinner.util.Constants.SPINNER_TITLE
+import com.example.luckyspinner.util.DialogUtil
 import com.example.luckyspinner.viewmodels.SpinnerListViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
@@ -141,6 +143,10 @@ class SpinnerListFragment : Fragment(), SpinnerListAdapter.Listener {
         binding.btnDoneAddChannel.setOnClickListener {
             lifecycleScope.launch(Dispatchers.Main) {
                 val nameSpinner = binding.edtEnterChannelName.text.toString()
+                if (nameSpinner == EMPTY_STRING) {
+                    binding.edtEnterChannelName.error = "Please fill this field"
+                    return@launch
+                }
                 val idSpinner = Calendar.getInstance().timeInMillis.toString()
                 viewModel.addSpinner(idChannel, Spinner(idSpinner, nameSpinner))
             }
@@ -161,7 +167,6 @@ class SpinnerListFragment : Fragment(), SpinnerListAdapter.Listener {
             it?.let {
                 if (it) {
                     Toast.makeText(context, "Edit Spinner Successfully!", Toast.LENGTH_SHORT).show()
-                    editSpinnerDiaLog.dismiss()
                 }
                 else {
                     Toast.makeText(context, "Edit Spinner Fail!!", Toast.LENGTH_SHORT).show()
@@ -175,7 +180,6 @@ class SpinnerListFragment : Fragment(), SpinnerListAdapter.Listener {
             it?.let {
                 if (it) {
                     Toast.makeText(context, "Delete Spinner Successfully!", Toast.LENGTH_SHORT).show()
-                    editSpinnerDiaLog.dismiss()
                 }
                 else {
                     Toast.makeText(context, "Delete Spinner Fail!!", Toast.LENGTH_SHORT).show()
@@ -233,7 +237,11 @@ class SpinnerListFragment : Fragment(), SpinnerListAdapter.Listener {
 
     override fun onDeleteItem(id: String) {
         lifecycleScope.launch(Dispatchers.Main) {
-            viewModel.deleteSpinner(idChannel, id)
+            val isDelete = DialogUtil.showYesNoDialog(context)
+            if (isDelete) {
+                progressDialog.show()
+                viewModel.deleteSpinner(idChannel, id)
+            }
         }
     }
